no sé que pasa con los tiempos

+++ src/philo.h
@@ -6,7 +6,7 @@
 /*   By: pmendez- <pmendez-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/24 20:04:41 by pmendez-          #+#    #+#             */
-/*   Updated: 2025/06/30 19:52:27 by pmendez-         ###   ########.fr       */
+/*   Updated: 2025/07/01 19:57:08 by pmendez-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -94,7 +94,7 @@ int		check_args(int argc, char **argv);
 // #############################################################################
 void	print_message_philo(t_philo *philo, char *message);
 int		ft_sleep(t_philo *philo, unsigned long total_sleep);
-size_t	get_time(void);
+unsigned long	get_time(void);
 int		check_if_philo_dead(t_philo *philo);
 
 // #############################################################################
